rename the file buildImageURL
