+++ app/lib/methods/helpers/buildImageURL.ts
